#121 structuring the modules/animal
