+++ app.ts
@@ -13,7 +13,8 @@ const app = express();
 app.use(applicationRateLimiter); // rate-limit applied to all the routes by default
 
 // Load Mock Modules
-require('./modules/animal/api/animal-routes')(app); // Animals
+require('./modules/address/address.routes')(app); // Addresses
+require('./modules/animal/animal.routes')(app); // Animals
 require('./modules/chat/api/chat-routes')(app); // Chat
 require('./modules/colors/api/colors-route')(app); // Colors
 require('./modules/countries/api/countries-routes')(app); // Countries
@@ -34,7 +35,6 @@ require('./modules/phone-numbers/api/phone-numbers-routes')(app); // Phone numbe
 require('./modules/quotes/api/quotes-routes')(app); // Quotes
 require('./modules/ip/api/ip-routes')(app); // ip and mac address
 require('./modules/vehicles/api/vehicles-routes')(app); // Vehicles
-require('./modules/address/address.routes')(app); // Addresses
 require('./modules/bankfeed/api/bankfeed-routes')(app); // Bank Feed
 require('./modules/location/api/location-routes')(app); // Bank Feed
 require('./modules/instruments/api/instruments-routes')(app); // Instruments
