+++ modules/animal/models/animal.types.ts
@@ -1,4 +1,4 @@
-enum AnimalType {
+enum Animal {
     Bear,
     Bird,
     Cat,
@@ -13,7 +13,7 @@ enum AnimalType {
     Rabbit,
     Rodent,
     Snake,
-    Type
+    Type,
 }
 
-export default AnimalType
+export default Animal;
